Make MINDS pipeline installable
