+++ src/minds/__init__.py
@@ -1,0 +1,6 @@
+from . import _version
+
+try:
+    __version__ = _version.version
+except Exception:
+    __version__ = "dev"
